feat: remake the comment feature and tag

+++ ProjectManagment/Application/TagsProjects/Commands/DeleteTagFromProjectCommand.cs
@@ -4,6 +4,7 @@ using Application.TagsProjects.Exceptions;
 using Domain.Projects;
 using Domain.Tags;
 using Domain.TagsProjects;
+using Domain.Users;
 using MediatR;
 
 namespace Application.TagsProjects.Commands;
@@ -11,10 +12,14 @@ namespace Application.TagsProjects.Commands;
 public class DeleteTagFromProjectCommand : IRequest<Result<TagsProject, TagProjectException>>
 {
     public required Guid ProjectId { get; init; }
+    public required Guid UserId { get; init; }
     public required Guid TagId { get; init; }
 }
 
-public class DeleteTagFromProjectCommandHandler(ITagProjectRepository repository)
+public class DeleteTagFromProjectCommandHandler(
+    ITagProjectRepository repository,
+    IUserRepository userRepository,
+    IProjectRepository projectRepository)
     : IRequestHandler<DeleteTagFromProjectCommand, Result<TagsProject, TagProjectException>>
 {
     public async Task<Result<TagsProject, TagProjectException>> Handle(DeleteTagFromProjectCommand request,
@@ -22,12 +27,33 @@ public class DeleteTagFromProjectCommandHandler(ITagProjectRepository repository
     {
         var projectId = new ProjectId(request.ProjectId);
         var tagId = new TagId(request.TagId);
-        var existingTagProject = await repository.GetByTagAndProjectId(tagId, projectId, cancellationToken);
+        var userId = new UserId(request.UserId);
 
-        return await existingTagProject.Match(
-            async tg => await DeleteEntity(tg, cancellationToken),
-            () => Task.FromResult<Result<TagsProject, TagProjectException>>(
-                new TagProjectNotFoundException(projectId, tagId)));
+        var existingTagProject = await repository.GetByTagAndProjectId(tagId, projectId, cancellationToken);
+        var existingUser = await userRepository.GetById(userId, cancellationToken);
+        var exstingProject = await projectRepository.GetById(projectId, cancellationToken);
+        return await existingUser.Match(
+            async u =>
+            {
+                return await exstingProject.Match(
+                    async p =>
+                    {
+                        if (p.UserId != userId || u.Role!.Name != "Admin")
+                        {
+                            return await Task.FromResult<Result<TagsProject, TagProjectException>>(
+                                new UserNotEnoughPremission(projectId, userId));
+                        }
+
+                        return await existingTagProject.Match(
+                            async tg => await DeleteEntity(tg, cancellationToken),
+                            () => Task.FromResult<Result<TagsProject, TagProjectException>>(
+                                new TagProjectNotFoundException(projectId, tagId)));
+                    },
+                    () => Task.FromResult<Result<TagsProject, TagProjectException>>(
+                        new ProjectNotFoundException(projectId))
+                );
+            },
+            () => Task.FromResult<Result<TagsProject, TagProjectException>>(new UserNotFound(projectId, userId)));
     }
 
     private async Task<Result<TagsProject, TagProjectException>> DeleteEntity(TagsProject tagsProject,
